refactor(core): split session strategies into function modules
