+++ packages/core/src/session/stateful/destroySession.ts
@@ -1,0 +1,82 @@
+import { getErrorName, verifyCSRFToken } from "@/shared/utils.ts"
+import type { InternalStatefulContext } from "@/@types/session.ts"
+
+export const __destroySession = ({ ctx, cookieConfig }: InternalStatefulContext) => {
+    const { logger, sessionConfig, cookies, jose } = ctx
+
+    return async (headers: Headers, skipCSRFCheck: boolean = false) => {
+        logger?.log("STATEFUL_DESTROY_SESSION_START", {
+            structuredData: {
+                strategy: "stateful",
+                operation: "destroySession",
+            },
+        })
+
+        await verifyCSRFToken({
+            headers,
+            cookies: cookies,
+            logger,
+            jose: jose,
+            skipCSRFCheck,
+        })
+
+        try {
+            const { sessionToken } = cookieConfig.getCookie(headers)
+            logger?.log("STATEFUL_SESSION_TOKEN_EXTRACTED", {
+                structuredData: {
+                    has_token: Boolean(sessionToken),
+                    token_length: sessionToken?.length || 0,
+                },
+            })
+
+            if (sessionToken) {
+                const sessionByToken = await sessionConfig.adapter.getSessionByToken(sessionToken)
+                logger?.log("STATEFUL_SESSION_DB_LOOKUP", {
+                    structuredData: {
+                        session_found: Boolean(sessionByToken),
+                        session_id: sessionByToken?.id || "",
+                    },
+                })
+
+                if (sessionByToken) {
+                    await sessionConfig.adapter.revokeSession(sessionByToken.id, "user_logout")
+                    logger?.log("STATEFUL_SESSION_REVOKED", {
+                        structuredData: {
+                            session_id: sessionByToken.id,
+                            reason: "user_logout",
+                        },
+                    })
+                } else {
+                    logger?.log("STATEFUL_SESSION_NOT_FOUND_FOR_DESTRUCTION", {
+                        structuredData: {
+                            reason: "session_not_found_in_database",
+                        },
+                    })
+                }
+            } else {
+                logger?.log("STATEFUL_NO_TOKEN_FOR_DESTRUCTION", {
+                    structuredData: {
+                        reason: "no_session_token_in_cookie",
+                    },
+                })
+            }
+        } catch (error) {
+            logger?.log("STATEFUL_DESTROY_SESSION_ERROR", {
+                structuredData: {
+                    error_type: getErrorName(error),
+                    error_message: error instanceof Error ? error.message : String(error),
+                },
+            })
+            throw error
+        }
+
+        const clearedHeaders = cookieConfig.clear()
+        logger?.log("STATEFUL_DESTROY_SESSION_SUCCESS", {
+            structuredData: {
+                cookies_cleared: true,
+            },
+        })
+
+        return clearedHeaders
+    }
+}
